requirements.txt
flask>=2.3
Pillow>=10.0
watchdog>=3.0
